SM: Moving common components to separate repository
