+++ src/main/java/com/bkk/sm/authentication/handler/impl/LoginHandlerImp.java
@@ -4,8 +4,8 @@ import com.bkk.sm.authentication.handler.LoginHandler;
 import com.bkk.sm.authentication.jwt.JWTUtil;
 import com.bkk.sm.authentication.model.AuthRequest;
 import com.bkk.sm.authentication.model.AuthResponse;
-import com.bkk.sm.authentication.model.User;
 import com.bkk.sm.authentication.services.mongo.UserDetailsService;
+import com.bkk.sm.mongo.authentication.model.User;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
