--- src/main/java/com/bkk/sm/authentication/model/Role.java
@@ -1,5 +1,0 @@
-package com.bkk.sm.authentication.model;
-
-public enum Role {
-    ROLE_SUPERADMIN, ROLE_USER, ROLE_ADMIN
-}
